+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## 0.0.2
+* allow configure glob pattern for the json and output file
 
-* TODO: Describe initial release.
+## 0.0.1
+* Run `dart run build_runner build` or run `flutter pub run build_runner build` to generate the class from json